Makefile: create target folders before using them
When installing to an "empty" filesystem (as is the case with fakeroots and
attempting to create installable packages), the ${PREFIX}/usr and ${PREFIX}/lib
directories need to be created (using "install -d").

Without the patch, if the filesystem is entirely empty, "libcsdr.so" gets
installed as "/lib" and "nmux" is installed as "/bin".